fix for Java 1.4
git-svn-id: http://svn.vuze.com/public/client@17394 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus3/src/com/aelitis/azureus/core/metasearch/utils/MomentsAgoDateFormatter.java
@@ -131,7 +131,7 @@ public class MomentsAgoDateFormatter {
            Integer field) {
        String result = null;
        long diff = now.getTimeInMillis() - then.getTimeInMillis();
-       long comparison = ((Integer)CONVERSION_MAP.get(field)).longValue();
+       long comparison = ((Long)CONVERSION_MAP.get(field)).longValue();
        if (diff > comparison) {
            long timeAgo = diff / comparison;
            result = String.valueOf(timeAgo).concat((String)UNIT_MAP.get(field));
